.env/requirements.txt
discord.py==1.7.3
dnspython==2.2.1
python-dotenv==0.20.0
